Default the icon_url to the bot's icon
When posting attachments, if the icon_url and icon_emoji fields aren't
specified, and the username is the same as the bot, pull the image from
the bot's user object and use that.

This doesn't fix the post coming from "Hubot (bot)", but at least it has
the right icon.